feat(step-flow): add new component
creates a new carbon component to aid consumers in creating and offering means
for end users to advance through a user journey. The component also has
built-in accessibility considerations with a `titleRef` prop being provided which
provides consumers with the ability to focus a visually hidden  DOM element
containing a formatted string with all of the relevant titles, category,
step progress contained within.

This ensures the component is accessible for all users, and is a more accessible
alternative to other step progress or sequence components.